fix warnings

+++ UIKit/Classes/UISegmentedControl.m
@@ -368,7 +368,7 @@ static NSString *kSSSegmentedControlEnabledKey = @"enabled";
         return nil;
     }
     
-    NSString *key = [NSString stringWithFormat:@"%i", index];
+    NSString *key = [NSString stringWithFormat:@"%i", (int)index];
     return [_segmentMeta objectForKey:key];
 }
 
@@ -393,7 +393,7 @@ static NSString *kSSSegmentedControlEnabledKey = @"enabled";
         _segmentMeta = [[NSMutableDictionary alloc] init];
     }
     
-    [_segmentMeta setValue:meta forKey:[NSString stringWithFormat:@"%i", index]];
+    [_segmentMeta setValue:meta forKey:[NSString stringWithFormat:@"%i", (int)index]];
     [self setNeedsDisplay];
 }
